Use std::optional for pin value
We have the need to express an invalid or empty pin value. We've
been doing this by setting to -1, however, there are some sensors
that output legitimate < 0 values, so this is no longer appropriate.
By using std::optional, we can have a value that can be set to
nullopt to indicate an invalid or empty value. This means we can
now utilise the entire integer space for sensor output.

This change is limited to the temp-netsender software,
we'll need to update the other clients at some point.