Allow building of weather netsender into temp-netsender
temp-netsender is becoming our all singing, all dancing external sensor
netsender client (soon to be renamed). This change will allows us to
optionally build in the weather sensor into the netsender client.

This is done by changing the SDLWeather class to implement the Sensor
interface, allowing us to add it to the polled sensor if we wish.