arduino: add pin safety functionality
This allows for compile time and runtime pin safety checks.
We also disallow use of pinMode in favour of using pinMode
wrappers that can validate mode.

When defining a static pin, we must acknowledge all of its
usage conditions. The intent here is that the programmer is
forced to be aware immediately of any caveats associated with
a pin.